// RetoDesarrolloFullStackPortales/frontend/src/app/models/client.model.ts
export class Client {
    id: Number;
    identificationType: String;
    identification: Number;
    firstName: String;
    lastName: String;
    email: String;
    birthdate: Date;
    creationDate: Date;
    userCreate: String;
    editDate: Date;
    userEdit: String;

    constructor(){
        this.id=new Number;
        this.identificationType=new String;
        this.identification=new Number;
        this.firstName=new String;
        this.lastName=new String;
        this.email=new String;
        this.birthdate=new Date;
        this.creationDate=new Date;
        this.userCreate=new String;
        this.editDate=new Date;
        this.userEdit=new String;
    }
}